Update: dev-demos/02 vegetable farm is working
Got something working in terms of stocking buildings like the
vegetable farm with supples as well as with resource consumption
by the people in the settlement.

Some buildings will earn restocking bonuses depending on what kind
of people are present in the settlement. For example, in this
current update, fruits and vegetables "grow faster" if there are
farmers present (although there are no farmers programmed in yet).